requirements.txt
httpx==0.27.2
psycopg[binary,pool]==3.2.10
python-dateutil==2.9.0.post0
tenacity==9.0.0
google-cloud-storage==2.18.2
boto3==1.34.162
python-dotenv==1.0.1
